requirements.txt
python-telegram-bot
requests
PySocks